Refactor AgentOps UI structure and mode persistence
