+++ src/agentops/lib/view.ts
@@ -1,0 +1,4 @@
+export function formatTime(value: string): string {
+  if (!value) return "-";
+  return value.replace("T", " ").replace("Z", " UTC");
+}
